[Simulator] Charts are empty in the screenshot
This is due to the combination of 2 factors:
* We resize the browser in order to take a screenshot with the real device size
* The sample redraws the charts on window resize (e.g.: pie-chart.js, ln 73)
As a result the screenshot is taken before the chart is rendered completely.

As a workaround we minimize the time for the chart redraw by disabling all transitions in it.
NB: The problem persists in stock-chart.js for some reason (data source too big?).

Related to: #301308
http://teampulse.telerik.com/view#item/301308